Add a simple API stub for fetching non-blocking flags.
We're going to move tracking of these properties to userspace, so that
we can provide them on a per-call basis, as opposed to having a single
flag per file descriptor.

This makes a lot more sense in an environment where file descriptor
passing is going to happen a lot. We don't want to be able to sabotage
processes by toggling their non-blocking flag.