Let userspace be the source of truth for O_NONBLOCK.
Change the value of O_NONBLOCK to be in the range of bits that are not
processed by the kernel. Instead, store these bits in the atomic integer
set we use for tracking non-blocked file descriptors.

For compatibility with existing kernels, attempt to keep the
kernel-space CLOUDABI_FDFLAG_NONBLOCK flag in sync with userspace as
much as possible.