First cut at implementing <uv.h>.
Now that we're going to remove kqueue and bring in an event loop, start
working on our own copy of libuv, which will support a subset of the
features provided by the official version that we can support.

We will first work towards implementing all of this on top of the plain
poll() syscall. This will allow us to temporarily remove all of
poll_fd() from CloudABI. Once all of that is done, we can sit around the
table and come up with a proper system call interface for optimizing
our own event loop.